% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_name_repair.R
\name{custom_name_repair}
\alias{custom_name_repair}
\title{Custom Name Repair}
\usage{
custom_name_repair(names)
}
\arguments{
\item{names}{Name of columns on dataset}
}
\value{
Vector of repaired names
}
\description{
Used in the survFMM function when only two subgroups are specified
}
