% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmm_em_algorithm.R
\name{fmm_em_algorithm}
\alias{fmm_em_algorithm}
\title{EM Algorithm for Finite Mixture Models with Survival Endpoints}
\usage{
fmm_em_algorithm(
  input_df,
  starting_values_input_df,
  k,
  outc_model_formula,
  outc_model_covars,
  weights_input,
  outc_distribution,
  covariates_subgroup_model,
  n_inits = 5,
  tolerance = 0.001,
  conv_pct_criteria = -1,
  max_iter = 200
)
}
\arguments{
\item{input_df}{Input data frame containing 1 row/observation, along with
each observation's event status (0=censored, 1=event) and event time
variables}

\item{starting_values_input_df}{Input dataset with starting values for algorithm}

\item{k}{Number of subgroups}

\item{weights_input}{Variable for inverse probability of treatment weights,
if applicable. For IPCW-FMM, supply the IPCW. For IPCW-FMM that also uses
IPTW, the product of the ITPW and IPCW may be supplied.}

\item{outc_distribution}{Outcome distribution for subgroup-specific outcome
models. Currently allowed values are "Weibull" and "Log-Normal" (not
case-sensitive)}

\item{covariates_subgroup_model}{Vector of covariates to include in subgroup membership model}

\item{n_inits}{Number of initial partitions for the EM algorithm. Default is
5. A higher number of initial partitions may result in greater stability of
estimates.}

\item{tolerance}{Convergence criteria for the change in the log-likelihood
for the EM algorithm.}

\item{conv_pct_criteria}{Convergence criteria for the percentage of
observations changing subgroup. Specify -1 to only use the log-likelihood
as the convergence criteria.}

\item{max_iter}{Maximum number of iterations. Default is 200.}
}
\value{
List
}
\description{
EM Algorithm for Finite Mixture Models with Survival Endpoints
}
