% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_test.R
\name{joint_test}
\alias{joint_test}
\title{Joint Test via Contrast Statement}
\usage{
joint_test(model = NULL, coef_df_input = NULL, cov_mtx_input = NULL, L, C)
}
\arguments{
\item{model}{Model object that contrast statement applies to}

\item{coef_df_input}{Coefficients for contrast}

\item{cov_mtx_input}{Covariance matrix for contrast}

\item{L}{Vector or matrix corresponding to contrast coefficients}

\item{C}{Vector of contrasts}
}
\value{
Tibble of contrast statement results containing the point estimate,
  test statistic, and p-value
}
\description{
Joint Test via Contrast Statement
}
