% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood_with_censoring.R
\name{log_likelihood_with_censoring}
\alias{log_likelihood_with_censoring}
\title{Log-likelihood computation for use within the survFMM function}
\usage{
log_likelihood_with_censoring(input_df, distribution)
}
\arguments{
\item{input_df}{Input data frame containing each observation's survival status (0=censored, 1=event) and time}

\item{distribution}{Acceptable values are currently Weibull or log-normal (not case sensitive)}
}
\value{
Log-likelihood value
}
\description{
Log-likelihood computation for use within the survFMM function
}
